refactor: cleanup dto
#10
