+++ src/events/dto/find-event.dto.ts
@@ -1,7 +1,22 @@
-import { ApiProperty, getSchemaPath } from '@nestjs/swagger';
-import { EventAttendeeDto } from './event-attendee.dto';
+import { ApiExtraModels, ApiProperty, getSchemaPath } from '@nestjs/swagger';
 import { EventDto } from './event.dto';
 
+export class EventAttendeeDto {
+  @ApiProperty({ description: '이벤트에 신청한 유저의 닉네임' })
+  intraId: string;
+
+  @ApiProperty({ description: '이벤트에 신청한 유저의 닉네임' })
+  url: string;
+
+  @ApiProperty({
+    type: Number,
+    description: '이벤트에 신청한 유저의 팀',
+    nullable: true,
+  })
+  teamId: number | null;
+}
+
+@ApiExtraModels(EventAttendeeDto)
 export class FindEventDto {
   @ApiProperty({ type: EventDto, description: '이벤트 정보' })
   event: EventDto;
